Likely going to move to a single very small hash block size instead of the tree format, still testing but overall performance seems to be better this way for now.

+++ inVtero.net/Mem.cs
@@ -169,12 +169,12 @@ namespace inVtero.net
             if (PTE.LargePage)
                 return false;
 
-            return DumpedPFNBitmap.GetBit(PTE.PFN);
+            return DumpedPFNBitmap.GetBit((ulong)PTE.PFN);
         }
 
         public void SetDumpedPFN(HARDWARE_ADDRESS_ENTRY PTE)
         {
-            DumpedPFNBitmap.SetBit(PTE.PFN);
+            DumpedPFNBitmap.SetBit((ulong)PTE.PFN);
         }
 
         public static Mem InitMem(String mFile, AMemoryRunDetector Detector, uint[] BitmapArray = null) //: this()
